+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/ModuleDeserializerStrategy.kt
@@ -1,19 +1,13 @@
 package com.insanusmokrassar.sdi
 
-import com.insanusmokrassar.sdi.utils.*
+import com.insanusmokrassar.sdi.utils.createModuleBasedOnConfigRoot
 import kotlinx.serialization.*
-import kotlinx.serialization.internal.*
+import kotlinx.serialization.internal.HashMapSerializer
+import kotlinx.serialization.internal.StringSerializer
 import kotlinx.serialization.json.*
-import kotlinx.serialization.modules.*
-import kotlin.reflect.KClass
 
 @ImplicitReflectionSerializer
-@Serializable(ModuleSerializer::class)
-class Module(base: Map<String, @ContextualSerialization Any>) : Map<String, Any> by base
-
-@ImplicitReflectionSerializer
-@Serializer(Module::class)
-internal object ModuleSerializer : KSerializer<Module> {
+internal object ModuleDeserializerStrategy : DeserializationStrategy<Module> {
     private val internalSerializer = HashMapSerializer(StringSerializer, ContextSerializer(Any::class))
     override val descriptor: SerialDescriptor
         get() = internalSerializer.descriptor
@@ -28,7 +22,5 @@ internal object ModuleSerializer : KSerializer<Module> {
         return Module(map)
     }
 
-    override fun serialize(encoder: Encoder, obj: Module) {
-        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-    }
+    override fun patch(decoder: Decoder, old: Module): Module = throw UpdateNotSupportedException("Module")
 }