+++ src/jvmMain/kotlin/com/insanusmokrassar/sdi/utils/currentSupertypes.kt
@@ -1,7 +1,0 @@
-package com.insanusmokrassar.sdi.utils
-
-import kotlin.reflect.KClass
-import kotlin.reflect.KType
-
-actual val KClass<*>.currentSupertypes: List<KType>
-    get() = this.supertypes