+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # SDI changelogs
 
+## 0.1.2
+
+* All reflection support methods are internal
+* Added `loadModule` functions for more useful working with modules loading
+* `Module` now is not serializable and can be created only via `loadModule` functions (at least, for some time)
+
 ## 0.1.1
 
 * Added opportunity to create objects inside of config: